Refactor configuration of clientcore
A service must be added via the add API method and unless no servies has
been added the application will exit immediately when start API method
is invoked.

Serveral services can be added but in order to keep the behaviour same
as before only the first service added is be used.